fix: move grant event model in separate file
